Change directory structure and add the test
