+++ public/assets/VueSchema.drawio
