+++ src/views/__test__/LoginView.test.js
